// src/app/modules/brand/brand.interface.ts
export type TBrand = {
  name: string;
  email: string;
  phone: string;
  productName: string;
  productLink: string;
  productType: string;
  tiktokHandle: string;
  tiktokLink: string;
  instragramHandle: string;
  instragramLink: string;
  websiteLink: string;
  additionalFormate: string;
  videoDuration: string;
  platForm: string;
  usageType: string;
  adHookOrCtaRequest?: string;
  exampleVideoLink: string;
  ugcPhotos: string[];
  ageRange: string;
  gender: string;
  location: string;
  language: string;
  script: string;
  anyWordsNotToUse: string;
  anySpecificWordsUse: string;
  howToPronouncebrandName: string;
  anySpecialRequest: string;
  expressDelivery: string;
  textOverlay?: string;
  captions?: string;
  music: string;
  extraHook: string;
  extraCta: string;
  videoType: string;
  additionalPerson: string;
  offSiteAttraction: string;
  goalOfProject: string;
  tergetAudience: string;
};